메인 페이지, 진행 중 프로젝트 수정
