+++ teaming/src/App.js
@@ -12,9 +12,9 @@ import FileDetail from "./pages/FileLinks/FileDetail";
 import End from "./pages/EndLinks/End";
 import { Calendar } from "./pages/Calendar";
 import {Home}  from "./pages/Home/Home.js";
-import { OngoingProject } from "./pages/OngoingProject";
 import FindPW from "./pages/LoginLinks/FindPW";
 import CleanHome from "./pages/CleanHome/CleanHome";
+import { OngoingProject } from "./pages/OngoingProject/OngoingProject";
 
 function App() {
   return (
